Refactored DeliteOps into low-level (IR), user-facing (DeliteOps) and Codegen files. Removed Foreach2 and ForeachBounded, deprecated ReduceTupleElem.
